+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 - New #342: Support JSON type (@Tigrov)
 - Chg #344: Change supported PHP versions to `8.1 - 8.4` (@Tigrov)
 - Chg #344: Change return type of `Command::insertWithReturningPks()` method to `array|false` (@Tigrov)
+- New #345: Add parameters `$ifExists` and `$cascade` to `CommandInterface::dropTable()` and
+  `DDLQueryBuilderInterface::dropTable()` methods (@vjik)
 
 ## 1.2.0 March 21, 2024